--- weekly-mission/signin/folder.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-    <head>
-      <meta charset="utf-8">
-      <title>Folder</title>
-      <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css" />
-      <link rel="preconnect" href="https://fonts.googleapis.com">
-      <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-      <link href="https://fonts.googleapis.com/css2?family=Acme&display=swap" rel="stylesheet">
-    </head>
-    <body>
-        folder
-    </body>
-</html>
