--- weekly-mission/signin/sign.css
@@ -1,99 +1,0 @@
-@import "../styles/color.css";
-@import "../styles/common.css";
-@import "../styles/ui.css";
-
-header {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  width: 218px;
-  height: 78px;
-}
-
-.logo-link {
-  display: flex;
-  justify-content: center;
-}
-
-.header-logo {
-  width: 210.58px;
-  height: 38px;
-}
-
-.header-message {
-  display: flex;
-  align-items: center;
-  width: 218px;
-  height: 24px;
-  column-gap: 8px;
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 24px;
-  color: #000000;
-}
-
-.header-link {
-  font-size: 16px;
-  font-weight: 600;
-  line-height: 19.09px;
-  color: var(--primary);
-}
-
-.sign-box {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 400px;
-  gap: 32px;
-}
-
-.eye-button {
-  position: absolute;
-  top: 45px;
-  right: 30px;
-  width: 30px;
-  height: 30px;
-  background-color: var(--background);
-}
-
-.login-button {
-  height: 53px;
-  padding: 16px 20px;
-  gap: 10px;
-  background-image: linear-gradient(90.99deg, var(--primary) 0.12%, #6AE3FE 101.84%);
-  color: #F5F5F5;
-  font-weight: 600;
-  font-size: 18px;
-  line-height: 21.48px;
-}
-
-.social-login_group {
-  display: flex;
-  width: 400px;
-  height: 66px;
-  padding: 12px 24px;
-  border-radius: 8px;
-  justify-content: space-between;
-  align-items: center;
-  background-color: var(--gray10);
-  border: 1px solid var(--gray20);
-}
-
-.social-login-text {
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 16.71px;
-  color: #373740;
-}
-
-.social-login-buttons {
-  width: 100px;
-  height: 42px;
-  gap: 16px;
-}
-
-.social-login-button {
-  width: 42px;
-  height: 42px;
-  border-radius: 50%;
-}
