--- weekly-mission/signin/signin.html
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>signin</title>
-    <link
-      rel="stylesheet"
-      type="text/css"
-      href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css"
-    />
-    <link rel="stylesheet" href="./sign.css" />
-  </head>
-  <body>
-    <div class="sign-container">
-      <header>
-        <a class="logo-link" href="../index.html">
-          <img class="header-logo" src="../image/Linkbrary.svg"/>
-        </a>
-        <p class="header-message">
-          회원이 아니신가요?<a class="header-link" href="../signup/signup.html">회원 가입하기</a>
-        </p>
-      </header>
-      <div class="sign-box">
-        <form class="sign-form">
-          <div class="sign-form-inputs">
-            <div class="sign-form-inputs-box">
-              <label class="sign-input-label">이메일</label>
-              <input class="email-input" type="email" />
-              <div class="error-message-email"></div>
-            </div>
-
-            <div class="sign-form-inputs-box sign-password">
-              <label class="sign-input-label">비밀번호</label>
-              <input class="password-input" type="password" />
-              <div class="error-message-password"></div>
-              <button class="eye-button" type="button">
-                <img src="../image/eye-off.jpg" />
-              </button>
-            </div>
-          </div>
-          <button class="login-button" type="submit">로그인</button>
-        </form>
-        <div class="social-login_group">
-          <span class="social-login-text">소셜 로그인</span>
-          <div class="social-login-buttons">
-            <a class="social-login-button google-link" href="https://www.google.com/">
-              <img src="../image/google icon.png" />
-            </a>
-            <a class="social-login-button kakao-link" href="https://www.kakaocorp.com/page/">
-              <img src="../image/kakao icon.png" />
-            </a>
-          </div>
-        </div>
-      </div>
-
-    <script src="./signin.js"></script>
-  </div>
-  </body>
-</html>
