--- weekly-mission/src/App.css
@@ -1,418 +1,0 @@
-@import "../styles/color.css";
-@import "../styles/common.css";
-@import "../styles/ui.css";
-
-/*body {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 0;
-}*/
-
-nav {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  position: sticky;
-  top: 0;
-  background-color: var(--background);
-  z-index: 2;
-}
-
-.gnb {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  height: 94px;
-  margin: 0;
-}
-
-.gnb-logo {
-  width: 133px;
-  height: 24px;
-  color: var(--primary);
-  font-weight: 800;
-  text-decoration-line: none;
-}
-
-.profile-info {
-  display: flex;
-  column-gap: 6px;
-}
-
-.gnb-button {
-  width: 128px;
-  height: 53px;
-  padding: 16px 20px;
-  background-image: linear-gradient(
-    90.99deg,
-    var(--primary) 0.12%,
-    #6ae3fe 101.84%
-  );
-  color: #f5f5f5;
-  font-size: 18px;
-  font-weight: 600;
-}
-
-header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  background-color: #edf7ff;
-  overflow: hidden;
-}
-
-.folder_info_container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding-top: 70px;
-  padding-bottom: 40px;
-  text-align: center;
-}
-
-.folder_info {
-  display: inline-block;
-  text-align: center;
-}
-
-.folder_info_profile {
-  margin-bottom: 20px;
-  width: 60px;
-  height: 60px;
-}
-
-.folder_info_owner {
-  display: block;
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 24px;
-}
-
-.folder_info_name {
-  margin-top: 20px;
-  font-size: 40px;
-  font-weight: 600;
-  line-height: 48px;
-  text-align: center;
-}
-
-.dim {
-  display: flex;
-  justify-content: center;
-  padding-top: 50px;
-}
-
-.dim-image {
-  position: relative;
-  top: 50px;
-  width: 1118px;
-  height: 659px;
-  border-radius: 25px;
-  background-color: var(--background);
-  box-shadow: 0 4px 25px #00000014;
-}
-
-article {
-  padding-top: 70px;
-  padding-bottom: 120px;
-}
-
-section {
-  display: grid;
-  justify-content: center;
-  width: 100%;
-  height: 550px;
-  column-gap: 157px;
-  padding: 50px 0;
-}
-
-section:nth-of-type(odd) {
-  grid-template:
-    ". image"
-    "title image"
-    "description image"
-    ". image"
-    /291px 550px;
-}
-
-section:nth-of-type(even) {
-  grid-template:
-    "image ."
-    "image title"
-    "image description"
-    "image ."
-    /550px 291px;
-}
-
-.section_title {
-  grid-area: title;
-  font-size: 48px;
-  font-weight: 700;
-  line-height: 57.6px;
-  margin: 0;
-}
-
-.section_text_description {
-  grid-area: description;
-  font-size: 16px;
-  font-weight: 400;
-  color: #6b6b6b;
-  line-height: 24px;
-}
-
-.section_img-box-image {
-  grid-area: image;
-  width: 550px;
-  height: 450px;
-}
-
-footer {
-  width: 100%;
-  height: 160px;
-  display: flex;
-  justify-content: center;
-  padding-top: 32px;
-  background-color: #111322;
-}
-
-.footerbox {
-  width: 100%;
-  height: fit-content;
-  display: flex;
-  justify-content: space-between;
-}
-
-.codeitlogo {
-  font-family: "Acme";
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 20px;
-  letter-spacing: 0em;
-  text-align: center;
-  color: #676767;
-}
-
-.info {
-  display: flex;
-  column-gap: 30px;
-  padding-right: 18px;
-}
-
-.icons {
-  display: flex;
-  column-gap: 12px;
-  height: 20px;
-}
-
-footer a {
-  gap: 30px;
-  font-family: "Acme";
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 20px;
-  letter-spacing: 0em;
-  text-align: left;
-  text-decoration-line: none;
-  color: #cfcfcf;
-}
-
-footer i {
-  color: var(--white);
-}
-
-/*태블릿 사이즈*/
-
-@media (max-width: 1199px) {
-  nav {
-    padding: 0 32px;
-  }
-
-  .gnb {
-    max-width: 800px;
-    padding: 0 32px;
-  }
-
-  header {
-    flex-direction: column;
-    column-gap: 40px;
-  }
-
-  .slogan {
-    font-size: 64px;
-    font-weight: 700;
-    width: 472px;
-    height: 240px;
-    line-height: 80px;
-  }
-
-  .goto-signup-link {
-    width: 350px;
-    height: auto;
-    padding: 16px 20px;
-    font-size: 18px;
-    line-height: 21.48px;
-  }
-
-  .dim {
-    padding: 29.07px 32px;
-  }
-
-  .dim-image {
-    width: 650.3px;
-    height: 383.11px;
-  }
-
-  section {
-    justify-content: center;
-    column-gap: 51px;
-    padding: 80px 251px;
-  }
-
-  .title {
-    font-size: 48px;
-    line-height: 57.28px;
-  }
-
-  .section_text_discription {
-    font-size: 16px;
-    line-height: 24px;
-  }
-
-  .section_img-box-image {
-    width: 385px;
-    height: 315px;
-  }
-
-  section:nth-of-type(odd) {
-    grid-template:
-      ". image"
-      "title image"
-      "description image"
-      ". image"
-      /262px 385px;
-  }
-
-  section:nth-of-type(even) {
-    grid-template:
-      "image ."
-      "image title"
-      "image description"
-      "image ."
-      /385px 262px;
-  }
-
-  .codeitlogo {
-    font-family: "Arial";
-  }
-
-  .info a {
-    font-family: "Arial";
-  }
-
-  .icons {
-    column-gap: 12px;
-  }
-}
-
-/*Mobile 사이즈*/
-
-@media (max-width: 769px) {
-  .gnb {
-    min-width: 400px;
-    padding: 0 32px;
-  }
-
-  header {
-    flex-direction: column;
-    column-gap: 24px;
-  }
-
-  .slogan {
-    font-size: 32px;
-    font-weight: 700;
-    width: 236px;
-    height: 126px;
-    line-height: 42px;
-  }
-
-  .goto-signup-link {
-    width: 200px;
-    height: auto;
-    padding: 10px 18px;
-    font-size: 14px;
-    line-height: 16.71px;
-  }
-
-  .dim {
-    padding: 13.56px 32px;
-  }
-
-  .dim-image {
-    width: 302.79px;
-    height: 178.71px;
-  }
-
-  section {
-    width: 100%;
-    height: auto;
-    display: flex;
-    flex-direction: column;
-    padding: 40px 32px;
-  }
-
-  .title {
-    font-size: 24px;
-    line-height: 28.64px;
-    display: inline;
-    gap: 10px;
-    margin-bottom: 24px;
-  }
-
-  .section_text_discription {
-    font-size: 14px;
-    line-height: 21px;
-  }
-
-  .section_img-box-image {
-    width: 100%;
-    height: auto;
-  }
-
-  footer {
-    width: 100%;
-    height: auto;
-    padding: 32px 32px;
-    margin: 0;
-  }
-
-  .footerbox {
-    display: grid;
-    grid-template:
-      "footer-info footer-icons"
-      "footer-logo .";
-    gap: 60px;
-    padding: 32px 32px;
-  }
-
-  .codeitlogo {
-    grid-area: footer-logo;
-    font-family: "Arial";
-  }
-
-  .info {
-    width: 181.11px;
-  }
-
-  .info a {
-    grid-area: footer-info;
-    font-family: "Arial";
-  }
-
-  .icons {
-    grid-area: footer-icons;
-    column-gap: 12px;
-  }
-}
