--- weekly-mission/styles/color.css
@@ -1,13 +1,0 @@
-:root {
-  --primary: #6d6afe;
-  --red: #ff5b56;
-  --black: #111322;
-  --white: #ffffff;
-
-  --gray100: #373740;
-  --gray60: #9fa6b2;
-  --gray20: #ccd5e3;
-  --gray10: #e7effb;
-
-  --background: #f0f6ff;
-}
