--- weekly-mission/styles/common.css
@@ -1,90 +1,0 @@
-* {
-  box-sizing: border-box;
-  font-family: "Pretendard";
-  margin: 0;
-}
-
-/*Basis styles*/
-
-button {
-  border: none;
-  border-radius: 8px;
-}
-
-/*landing styles*/
-
-.gnb {
-  max-width: 1920px;
-  padding: 0 200px;
-}
-
-.footerbox {
-  max-width: 1920px;
-  padding: 0 104px;
-}
-
-
-@media (max-width: 1199px) {
-  .gnb {
-    max-width: 800px;
-    padding: 0 32px;
-  }
-}
-
-@media (max-width: 769px) {
-  .gnb {
-    min-width: 400px;
-    padding: 0 32px;
-  }
-
-  .footerbox {
-    padding: 32px 32px;
-  }
-}
-
-/*sign styles*/
-
-.sign-container {
-  padding-top: 238px;
-  background-color: var(--background);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  row-gap: 30px;
-}
-
-.sign-form {
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-  width: 100%;
-}
-
-.sign-form-inputs {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-}
-
-.sign-form-inputs-box {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.sign-password {
-  position: relative;
-}
-
-@media (max-width: 1199px) {
-  .sign-container {
-    padding-top: 200px;
-  }
-}
-
-@media (max-width: 769px) {
-  .sign-container {
-    padding-top: 120px;
-  }
-}
-
