--- weekly-mission/styles/ui.css
@@ -1,73 +1,0 @@
-/*landingopage styling*/
-#maincontents-text-purple-to-pink {
-  background-clip: text;
-  -webkit-background-clip: text;
-  color: transparent;
-  background-image: linear-gradient(91.26deg, var(--primary) 17.28%, #FF9F9F 74.98%);
-}
-
-#section_text-color-pink-to-blue {
-  background-clip: text;
-  -webkit-background-clip: text;
-  color: transparent;
-  background-image: linear-gradient(95.63deg, #FE8A8A 1.72%, #A4CEFF 74.97%);
-}
-
-#section_text-color-yellow-to-blue {
-  background-clip: text;
-  -webkit-background-clip: text;
-  color: transparent;
-  background-image: linear-gradient(277.01deg, #6FBAFF 59.22%, #FFD88B 93.66%);
-}
-
-#section_text-color-purple-to-blue {
-  background-clip: text;
-  -webkit-background-clip: text;
-  color: transparent;
-  background-image: linear-gradient(99.5deg, #6D7CCD 19.76%, rgba(82, 136, 133, 0.22) 52.69%);  
-}
-
-#section_text-color-blue-to-purple {
-  background-clip: text;
-  -webkit-background-clip: text;
-  color: transparent;
-  background-image: linear-gradient(271.36deg, #FE578F -9.84%, #68E8F9 107.18%);
-}
-
-/*signpage styling*/
-.sign-input-label {
-  font-size: 14px;
-  font-weight: 400;
-  line-height: 17px;
-  letter-spacing: 0em;
-  text-align: left;
-  color: #000000;
-}
-
-input {
-  width: 400px;
-  height: 60px;
-  padding: 18px, 15px;
-  border-radius: 8px;
-  border: 1px solid var(--primary);
-  outline: none;
-}
-
-input:focus {
-  outline: 1px solid blue;
-  border: none;
-}
-
-.wrongsign {
-  border: none;
-  outline: 1px solid red;
-}
-
-.error-message-email,
-.error-message-password {
-  display: block;
-  font-weight: 400;
-  font-size: 14px;
-  color: #FF5B56;
-  line-height: 16.71px;
-}
